Fixed #28: Reload binary
Fixed checkSave() fall through on save-as that is canceled

If project is modified and checkSave() calls saveProject() and the dialog was canceled then
checkSave() would still return true. That means that in cases where checkSave() is called the code
could continue without saving the project first, which is not very user-friendly.

Check binary save before reloading binary